+++ src/ahk/constants.js
@@ -1,1 +1,1 @@
-export const AHKGEN_SIGNATURE_PREFIX = "; Made with AHKgen";
+export const AHKFORGE_SIGNATURE_PREFIX = "; Made with AHKforge";